+++ NEWS.md
@@ -4,6 +4,11 @@
   by default; use `just = 0`/`just = 1` to place columns on the left/right
   of the axis breaks.
   (@wurli, #4899)
+  
+* Improve the support for `guide_axis()` on `coord_trans()` (@yutannihilation, #3959)
+
+* `geom_density()` and `stat_density()` now support `bounds` argument
+  to estimate density with boundary correction (@echasnovski, #4013).
 
 * ggplot now checks during statistical transformations whether any data 
   columns were dropped and warns about this. If stats intend to drop